CLIENT - GroupButton included on the shared component
