+++ client/src/shared-components/index.js
@@ -1,7 +1,7 @@
 import {PrimaryButton as PB} from './Button/PrimaryButton';
 import {SecondaryButton as SB} from './Button/SecondaryButton';
 import {TertiaryButton as TB} from './Button/TertiaryButton';
-
+import {GroupButton as GB} from './Button/GroupButton';
 import {SpanAlert as SA} from './Span/SpanAlert';
 import {Stars as S} from './Stars/Stars';
 import {Tag as T} from './Tag/Tag'
@@ -10,6 +10,7 @@ import {theme as t} from './theme';
 export const PrimaryButton = PB;
 export const SecondaryButton = SB;
 export const TertiaryButton = TB;
+export const GroupButton = GB;
 export const SpanAlert = SA;
 export const Stars = S;
 export const Tag = T;
